remove references to long, incompatible with python3

+++ zcrmsdk/Operations.py
@@ -551,7 +551,7 @@ class ZCRMJunctionRecord(object):
 class ZCRMLeadConvertMapping(object):
     
     def __init__(self,name,converted_id):
-        self.id=long(converted_id)
+        self.id=converted_id
         self.name=name
         self.fields=list()
     @staticmethod
@@ -600,7 +600,7 @@ class ZCRMModuleRelatedList(object):
         return ZCRMModuleRelatedList(api_name)
     
     def set_relatedlist_properties(self,relatedlist_prop):
-        self.id=long(relatedlist_prop['id'])
+        self.id=relatedlist_prop['id']
         self.module=relatedlist_prop['module']
         self.display_label=relatedlist_prop['display_label']
         self.name=relatedlist_prop['name']
